Move code to new util lib
