+++ libs/dh/shared/data-access-graphql/src/index.ts
@@ -18,12 +18,3 @@ export { graphQLProviders } from './lib/dh-graphql.providers';
 export { parseGraphQLErrorResponse } from './lib/error-handler';
 export { getActorOptions } from './lib/graphql/get-actors-by-eic-function-for-dropdowns';
 export { getGridAreaOptions } from './lib/graphql/get-grid-area-options';
-export {
-  query,
-  lazyQuery,
-  mutation,
-  LazyQueryOptions,
-  MutationOptions,
-  QueryOptions,
-  SubscribeToMoreOptions,
-} from './lib/signal';
