+++ libs/dh/shared/util-apollo/project.json
@@ -1,0 +1,20 @@
+{
+  "name": "dh-shared-util-apollo",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/dh/shared/util-apollo/src",
+  "prefix": "dh",
+  "projectType": "library",
+  "tags": ["product:dh", "domain:shared", "type:util"],
+  "targets": {
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "libs/dh/shared/util-apollo/jest.config.ts"
+      }
+    },
+    "lint": {
+      "executor": "@nx/eslint:lint"
+    }
+  }
+}
