+++ libs/dh/shared/util-apollo/src/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/query';
+export * from './lib/lazyQuery';
+export * from './lib/mutation';
