+++ libs/dh/shared/util-apollo/src/lib/query.ts
@@ -37,11 +37,9 @@ import {
   startWith,
   switchMap,
   take,
-  tap,
   withLatestFrom,
 } from 'rxjs';
 import { exists } from '@energinet-datahub/dh/shared/util-operators';
-import { MutationOptions as ApolloMutationOptions } from 'apollo-angular/types';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 
 /** Create an observable of ApolloQueryResult from an ApolloError. */
@@ -176,99 +174,6 @@ export function query<TResult, TVariables extends OperationVariables>(
 
       // Allow the caller to terminate the subscription manually
       return () => subscription.unsubscribe();
-    },
-  };
-}
-
-// Lazy queries cannot be skipped as they are triggered imperatively
-export interface LazyQueryOptions<TResult, TVariables extends OperationVariables>
-  extends Omit<QueryOptions<TVariables>, 'skip'> {
-  onCompleted?: (data: TResult) => void;
-  onError?: (error: ApolloError) => void;
-}
-
-/** Signal-based wrapper around Apollo's `query` function, made to align with `useLazyQuery`. */
-export function lazyQuery<TResult, TVariables extends OperationVariables>(
-  // Limited to TypedDocumentNode to ensure the query is statically typed
-  document: TypedDocumentNode<TResult, TVariables>,
-  options?: LazyQueryOptions<TResult, TVariables>
-) {
-  const result = query(document, { ...options, skip: true });
-  return {
-    ...result,
-    query: (options: Partial<LazyQueryOptions<TResult, TVariables>>) =>
-      result.setOptions(options).then((result) => {
-        if (result.error) {
-          options?.onError?.(result.error);
-        } else if (result.data) {
-          options?.onCompleted?.(result.data);
-        }
-
-        // Only tap into the chain for side-effects
-        return result;
-      }),
-  };
-}
-
-// Add the `onCompleted` and `onError` callbacks to align with `useMutation`
-export interface MutationOptions<TResult, TVariables>
-  extends Omit<ApolloMutationOptions<TResult, TVariables>, 'mutation'> {
-  onCompleted?: (data: TResult, clientOptions?: MutationOptions<TResult, TVariables>) => void;
-  onError?: (error: ApolloError, clientOptions?: MutationOptions<TResult, TVariables>) => void;
-}
-
-/** Signal-based wrapper around Apollo's `mutate` function, made to align with `useMutation`. */
-export function mutation<TResult, TVariables>(
-  // Limited to TypedDocumentNode to ensure the query is statically typed
-  document: TypedDocumentNode<TResult, TVariables>,
-  options?: MutationOptions<TResult, TVariables>
-) {
-  // Rename the options to avoid shadowing
-  const parentOptions = options;
-
-  // Inject dependencies
-  const client = inject(Apollo);
-  const destroyRef = inject(DestroyRef);
-
-  // Signals holding the result values
-  const data = signal<TResult | undefined>(undefined);
-  const error = signal<ApolloError | undefined>(undefined);
-  const loading = signal(true);
-
-  return {
-    // Upcast to prevent writing to signals
-    data: data as Signal<TResult | undefined>,
-    error: error as Signal<ApolloError | undefined>,
-    loading: loading as Signal<boolean>,
-    mutate(options: MutationOptions<TResult, TVariables>) {
-      const mergedOptions = { ...parentOptions, ...options };
-      const { onCompleted, onError, ...mutationOptions } = mergedOptions;
-      return firstValueFrom(
-        client.mutate({ ...mutationOptions, mutation: document }).pipe(
-          // The MutationResult type is different from QueryResult in several ways
-          map(({ errors, ...result }) => ({
-            ...result,
-            error: errors?.length ? new ApolloError({ graphQLErrors: errors }) : undefined,
-          })),
-          catchError((error: ApolloError) => of({ error, data: undefined, loading: false })),
-          tap((result) => {
-            data.set(result.data ?? undefined);
-            error.set(result.error);
-            loading.set(result.loading);
-          }),
-          // Since this observable returns a promise, it should only emit the final result
-          filter((result) => !result.loading),
-          tap((result) => {
-            if (result.error) {
-              onError?.(result.error, mergedOptions);
-            } else if (result.data) {
-              onCompleted?.(result.data, mergedOptions);
-            }
-          }),
-          take(1), // Complete the observable when result is available
-          takeUntilDestroyed(destroyRef) // Or when the component is destroyed
-        )
-      );
     },
   };
 }
