+++ libs/dh/shared/util-apollo/src/test-setup.ts
@@ -1,0 +1,5 @@
+import 'jest-preset-angular/setup-jest';
+
+import { setUpTestbed } from '@energinet-datahub/gf/test-util-staging';
+
+setUpTestbed();
